Update FetchMSIProperties.VBS
For just 1 or 2 MSIs, we usually do this task by using orca or installshield , However for large packages such as AutoCAD or Microsoft or SAP etc., where there are more then 10-15 MSIs, fetching the properties manually becomes quite a lengthy and exhaustive process. This script is created to help you in such situation.